Regras de negocio

+++ back-end/src/main/java/com/basis/srs/servico/EquipamentoServico.java
@@ -1,7 +1,9 @@
 package com.basis.srs.servico;
 
 import com.basis.srs.dominio.Equipamento;
+import com.basis.srs.exception.RegraNegocioException;
 import com.basis.srs.repositorio.EquipamentoRepositorio;
+import com.basis.srs.repositorio.SalaEquipamentoRepositorio;
 import com.basis.srs.servico.dto.EquipamentoDTO;
 import com.basis.srs.servico.mapper.EquipamentoMapper;
 import lombok.RequiredArgsConstructor;
@@ -17,6 +19,8 @@ public class EquipamentoServico {
 
     private final EquipamentoRepositorio equipamentoRepositorio;
 
+    private final SalaEquipamentoRepositorio salaEquipamentoRepositorio;
+
     private final EquipamentoMapper equipamentoMapper;
 
     public List<EquipamentoDTO> listar(){
@@ -28,7 +32,7 @@ public class EquipamentoServico {
     };
 
     public EquipamentoDTO buscarPorId(Integer id){
-        Equipamento equipamento = equipamentoRepositorio.findById(id).orElse(null);
+        Equipamento equipamento = equipamentoRepositorio.findById(id).orElseThrow(()->new RegraNegocioException("Equipamento não ecnotrnado"));
 
         return equipamentoMapper.toDto(equipamento);
     };
@@ -43,6 +47,11 @@ public class EquipamentoServico {
 
 
     public void deletar(Integer id){
+
+        if(salaEquipamentoRepositorio.existsByEquipamentoId(id)){
+            throw new RegraNegocioException("Equipamento esta cadastrado em uma sala");
+        }
+
         equipamentoRepositorio.deleteById(id);
     };
